This whole app seems to be unneeded
It may even have been that it was already obsolete when it was created, I just didn't know it. Either way, it seems that GitHub may be even smarter than I am ;-).

One thing that is worth noting that deleting a branch via git CLI does *not* trigger this behavior.